db/requirements.txt
# Arc-Halo Cognitive Fusion Reactor - Python Dependencies
# Install with: pip install -r requirements.txt

# Database connectivity
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Environment management
python-dotenv>=1.0.0

# Tensor operations (for future integration)
numpy>=1.24.0

# Logging and monitoring
structlog>=23.2.0

# Data validation
pydantic>=2.5.0

# Optional: Neon API client
requests>=2.31.0
